requirements.txt
matplotlib
Pillow
opsdroid-get-image-size